+++ tests-integration/tests/req_test/merge.rs
@@ -27,16 +27,16 @@ use googletest::prelude::near;
 
 #[test]
 fn merge_into_empty_preserves_source_distribution() {
-    let mut target: ReqSketch<f32> = ReqSketch::new(40, RankAccuracy::HighRank).unwrap();
-    let mut source: ReqSketch<f32> = ReqSketch::new(40, RankAccuracy::HighRank).unwrap();
+    let mut target: ReqSketch<i32> = ReqSketch::new(40, RankAccuracy::HighRank).unwrap();
+    let mut source: ReqSketch<i32> = ReqSketch::new(40, RankAccuracy::HighRank).unwrap();
 
     for i in 0..1000 {
-        source.update(i as f32);
+        source.update(i);
     }
 
     target.merge(&source).expect("merge should succeed");
-    assert_eq!(target.min_item(), Some(&0.0));
-    assert_eq!(target.max_item(), Some(&999.0));
+    assert_eq!(target.min_item(), Some(&0));
+    assert_eq!(target.max_item(), Some(&999));
 
     let q25 = target
         .quantile(0.25, SearchCriteria::Inclusive)
@@ -48,30 +48,30 @@ fn merge_into_empty_preserves_source_distribution() {
         .quantile(0.75, SearchCriteria::Inclusive)
         .expect("quantile should succeed");
     let r50 = target
-        .rank(&500.0, SearchCriteria::Inclusive)
+        .rank(&500, SearchCriteria::Inclusive)
         .expect("rank should succeed");
 
-    assert_that!(q25, near(250.0, 250.0 * 0.01));
-    assert_that!(q50, near(500.0, 500.0 * 0.01));
-    assert_that!(q75, near(750.0, 750.0 * 0.01));
+    assert_that!(q25 as f64, near(250.0, 250.0 * 0.01));
+    assert_that!(q50 as f64, near(500.0, 500.0 * 0.01));
+    assert_that!(q75 as f64, near(750.0, 750.0 * 0.01));
     assert_that!(r50, near(0.5, 0.5 * 0.01));
 }
 
 #[test]
 fn merge_two_ranges_preserves_distribution() {
-    let mut left: ReqSketch<f32> = ReqSketch::new(100, RankAccuracy::HighRank).unwrap();
-    let mut right: ReqSketch<f32> = ReqSketch::new(100, RankAccuracy::HighRank).unwrap();
+    let mut left: ReqSketch<i32> = ReqSketch::new(100, RankAccuracy::HighRank).unwrap();
+    let mut right: ReqSketch<i32> = ReqSketch::new(100, RankAccuracy::HighRank).unwrap();
 
     for i in 0..1000 {
-        left.update(i as f32);
+        left.update(i);
     }
     for i in 1000..2000 {
-        right.update(i as f32);
+        right.update(i);
     }
 
     left.merge(&right).expect("merge should succeed");
-    assert_eq!(left.min_item(), Some(&0.0));
-    assert_eq!(left.max_item(), Some(&1999.0));
+    assert_eq!(left.min_item(), Some(&0));
+    assert_eq!(left.max_item(), Some(&1999));
 
     let q25 = left
         .quantile(0.25, SearchCriteria::Inclusive)
@@ -83,21 +83,21 @@ fn merge_two_ranges_preserves_distribution() {
         .quantile(0.75, SearchCriteria::Inclusive)
         .expect("quantile should succeed");
     let r50 = left
-        .rank(&1000.0, SearchCriteria::Inclusive)
+        .rank(&1000, SearchCriteria::Inclusive)
         .expect("rank should succeed");
 
-    assert_that!(q25, near(500.0, 500.0 * 0.02));
-    assert_that!(q50, near(1000.0, 1000.0 * 0.01));
-    assert_that!(q75, near(1500.0, 1500.0 * 0.01));
+    assert_that!(q25 as f64, near(500.0, 500.0 * 0.02));
+    assert_that!(q50 as f64, near(1000.0, 1000.0 * 0.01));
+    assert_that!(q75 as f64, near(1500.0, 1500.0 * 0.01));
     assert_that!(r50, near(0.5, 0.5 * 0.01));
 }
 
 #[test]
 fn merge_rejects_incompatible_accuracy_modes() {
     let mut high_rank = ReqSketch::default();
-    let low_rank: ReqSketch<f32> = ReqSketch::new(12, RankAccuracy::LowRank).unwrap();
+    let low_rank: ReqSketch<i32> = ReqSketch::new(12, RankAccuracy::LowRank).unwrap();
 
-    high_rank.update(1.0);
+    high_rank.update(1);
     assert_that!(high_rank.merge(&low_rank), err(anything()));
 }
 
@@ -108,17 +108,17 @@ fn many_small_merges_preserve_count_bounds_and_median() {
     for batch in 0..100 {
         let mut batch_sketch = ReqSketch::default();
         for i in 0..100 {
-            batch_sketch.update((batch * 100 + i) as f64);
+            batch_sketch.update(batch * 100 + i);
         }
         sketch.merge(&batch_sketch).expect("merge should succeed");
     }
 
     assert_eq!(sketch.n(), 10_000);
-    assert_eq!(sketch.min_item(), Some(&0.0));
-    assert_eq!(sketch.max_item(), Some(&9999.0));
+    assert_eq!(sketch.min_item(), Some(&0));
+    assert_eq!(sketch.max_item(), Some(&9999));
 
     let median = sketch
         .quantile(0.5, SearchCriteria::Inclusive)
         .expect("quantile should succeed");
-    assert_that!(median, near(4999.5, 500.0));
+    assert_that!(median as f64, near(4999.5, 500.0));
 }
