+++ tests-integration/tests/req_test/structure.rs
@@ -30,7 +30,7 @@ use googletest::prelude::predicate;
 fn iterator_weights_sum_to_n_and_items_are_in_range() {
     let mut sketch = ReqSketch::default();
     for i in 0..1000 {
-        sketch.update(i as f64);
+        sketch.update(i);
     }
 
     let total_weight: u64 = sketch.iter().map(|(_, weight)| weight).sum();
@@ -48,10 +48,10 @@ fn small_sketch_iterator_reports_unit_weights() {
     // Below the compaction threshold every retained item still has weight 1.
     let mut sketch = ReqSketch::default();
     for i in 0..10 {
-        sketch.update(i as f64);
+        sketch.update(i);
     }
 
-    let items: Vec<(f64, u64)> = sketch.iter().collect();
+    let items: Vec<(i32, u64)> = sketch.iter().collect();
     assert_eq!(items.len(), 10);
     let weights: Vec<_> = items.iter().map(|&(_, weight)| weight).collect();
     assert_that!(weights, each(eq(&1)));
@@ -69,7 +69,7 @@ fn compaction_promotes_surviving_items_to_higher_weights() {
     // level at double weight, so the maximum item weight exceeds 1.
     let mut sketch = ReqSketch::default();
     for i in 0..100_000 {
-        sketch.update(i as f64);
+        sketch.update(i);
     }
 
     let max_weight = sketch.iter().map(|(_, weight)| weight).max().unwrap();
